move events.json under assets/data
