+++ src/app/events/[id]/page.tsx
@@ -4,7 +4,7 @@ import * as S from './page.styled';
 import { useParams } from 'next/navigation';
 import { getFormattedDate, getFormattedTime } from '@/utils/formatUtil';
 import DefaultButton from '@/components/Button/DefaultButton';
-import eventSource from '@/assets/events.json';
+import eventSource from '@/assets/data/events.json';
 
 export default function EventInfo() {
   const { id } = useParams();
