+++ src/app/events/past/page.tsx
@@ -3,7 +3,7 @@
 import * as S from './page.styled';
 import { getFormattedDate } from '@/utils/formatUtil';
 import CEventItem from '@/components/c-event-item';
-import eventSource from '@/assets/events.json';
+import eventSource from '@/assets/data/events.json';
 
 export default function Events() {
   const pastEvents = eventSource.events;
